Move leaderboard.py into utils folder
